working with smash
