+++ training_mod_tui/src/main.rs
@@ -27,70 +27,15 @@ fn main() -> Result<(), Box<dyn Error>> {
         menu = get_menu();
     }
 
-    let tab_specifiers = vec![
-        ("Mash Settings", vec![
-            "Mash Toggles",
-            "Followup Toggles",
-            "Attack Angle",
-            "Ledge Options",
-            "Ledge Delay",
-            "Tech Options",
-            "Miss Tech Options",
-            "Defensive Options",
-            "Aerial Delay",
-            "OoS Offset",
-            "Reaction Time",
-        ]),
-        ("Defensive Settings", vec![
-            "Fast Fall",
-            "Fast Fall Delay",
-            "Falling Aerials",
-            "Full Hop",
-            "Shield Tilt",
-            "DI Direction",
-            "SDI Direction",
-            "Airdodge Direction",
-            "SDI Strength",
-            "Shield Toggles",
-            "Mirroring",
-            "Throw Options",
-            "Throw Delay",
-            "Pummel Delay",
-            "Buff Options",
-        ]),
-        ("Miscellaneous Settings", vec![
-            "Input Delay",
-            "Save States",
-            "Save Damage",
-            "Hitbox Visualization",
-            "Stage Hazards",
-            "Frame Advantage",
-            "Mash In Neutral"
-        ])
-    ];
-    let mut tabs: std::collections::HashMap<&str, Vec<SubMenu>> = std::collections::HashMap::new();
-    tabs.insert("Mash Settings", vec![]);
-    tabs.insert("Defensive Settings", vec![]);
-    tabs.insert("Miscellaneous Settings", vec![]);
-
-    let mut menu_items = Vec::new();
-    for sub_menu in menu.sub_menus.iter() {
-        for tab_spec in tab_specifiers.iter() {
-            if tab_spec.1.contains(&sub_menu.title) {
-                tabs.get_mut(tab_spec.0).unwrap().push(sub_menu.clone());
-                menu_items.push( sub_menu.clone());
-            }
-        }
-    };
-
-    let app = training_mod_tui::App::new(tabs, menu_items, 3);
+    let mut app = training_mod_tui::App::new(menu);
 
     #[cfg(not(feature = "has_terminal"))] {
-        let backend = tui::TestBackend::new(100, 8);
+        let backend = tui::backend::TestBackend::new(100, 8);
         let mut terminal = Terminal::new(backend)?;
         let mut state = ListState::default();
         state.select(Some(1));
-        let frame_res = terminal.draw(|f| training_mod_tui::ui(f, &mut app))?;
+        let mut url = String::new();
+        let frame_res = terminal.draw(|f| url = training_mod_tui::ui(f, &mut app))?;
 
         for (i, cell) in frame_res.buffer.content().into_iter().enumerate() {
             print!("{}", cell.symbol);
@@ -123,6 +68,8 @@ fn main() -> Result<(), Box<dyn Error>> {
 
         if let Err(err) = res {
             println!("{:?}", err)
+        } else {
+            println!("URL: {}", res.as_ref().unwrap());
         }
     }
 
@@ -133,10 +80,11 @@ fn run_app<B: Backend>(
     terminal: &mut Terminal<B>,
     mut app: training_mod_tui::App,
     tick_rate: Duration,
-) -> io::Result<()> {
+) -> io::Result<String> {
     let mut last_tick = Instant::now();
+    let mut url = String::new();
     loop {
-        terminal.draw(|f| training_mod_tui::ui(f, &mut app))?;
+        terminal.draw(|f| url = training_mod_tui::ui(f, &mut app).clone())?;
 
         let timeout = tick_rate
             .checked_sub(last_tick.elapsed())
@@ -146,7 +94,7 @@ fn run_app<B: Backend>(
         if crossterm::event::poll(timeout)? {
             if let Event::Key(key) = event::read()? {
                 match key.code {
-                    KeyCode::Char('q') => return Ok(()),
+                    KeyCode::Char('q') => return Ok((url)),
                     KeyCode::Char('r') => app.on_r(),
                     KeyCode::Char('l') => app.on_l(),
                     KeyCode::Left => app.on_left(),
